Implemented redux-persist
