+++ package.json
@@ -41,6 +41,7 @@
     "redux-devtools-dock-monitor": "^1.1.2",
     "redux-devtools-log-monitor": "^1.3.0",
     "redux-logger": "^3.0.6",
+    "redux-persist": "^5.4.0",
     "redux-promise": "^0.5.3",
     "redux-storage": "^4.1.2",
     "redux-thunk": "^2.2.0",
